+++ examples/__init__.py
@@ -1,4 +1,0 @@
-"""
-Package marker for bundled examples.
-"""
-